2025.1 Release

+++ DirectProgramming/C++SYCL/Jupyter/gpu-optimization-sycl-training/06_Shared_Local_Memory/lab/convolution_slm.cpp
@@ -6,8 +6,8 @@
 #include <sycl/sycl.hpp>
 
 int main() {
-  constexpr size_t N = 8192 * 8192;
-  constexpr size_t M = 257;
+  constexpr int N = 8192 * 8192;
+  constexpr int M = 257;
 
   std::vector<int> input(N);
   std::vector<int> output(N);
